+++ packages/components/typography/Text.tsx
@@ -82,7 +82,7 @@ const Text = forwardRef<HTMLSpanElement, TypographyTextProps>((originalProps, re
         }
       : {
           text: copyable?.text || children?.toString(),
-          onCopy: copyable?.onCopy?.(),
+          onCopy: copyable?.onCopy,
           tooltipProps: {
             ...copyable?.tooltipProps,
             content: isCopied ? copiedText : copyable?.tooltipProps?.content,
@@ -102,7 +102,13 @@ const Text = forwardRef<HTMLSpanElement, TypographyTextProps>((originalProps, re
   const renderContent = (withChildren: boolean) => {
     const { tooltipProps } = copyProps;
     const wrapWithTooltip = (wrapContent: React.ReactNode) =>
-      tooltipProps ? <Tooltip {...tooltipProps}>{wrapContent}</Tooltip> : wrapContent;
+      tooltipProps ? (
+        <Tooltip hideEmptyPopup {...tooltipProps}>
+          {wrapContent}
+        </Tooltip>
+      ) : (
+        wrapContent
+      );
 
     const getSuffix = (): ReactElement => {
       if (typeof copyProps?.suffix === 'function') {